Rename tool target to same as project. Boom! The scheme is now borked. This is phase 1 of the bug.
The scheme has now been modified by Xcode to contain a reference to build the tool, and the user cannot delete. That is a bug, but I haven't yet fully duplicated the steps to make Xcode start RANDOMLY corrupting the scheme, or not, on each launch. (That makes the bug way worse because the user can lose a whole day debugging it, since whatever they do may seem to work on any given launch, but really have no effect.) So I am gonna keep plugging to see if I can reproduce THAT aspect.